requirements.txt
fastapi==0.115.6
psycopg2==2.9.10
SQLAlchemy==2.0.36
uvicorn==0.32.1
pydantic-settings==2.6.1
asyncpg==0.30.0
email_validator==2.2.0
pytest==8.3.4
httpx==0.28.1
sqlmodel==0.0.22
fastapi-restful[all]==0.6.0
fastapi-filter==2.0.1
